project_listview/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.txt
#
asgiref==3.2.10           # via -r requirements.txt, django
django==3.1               # via -r requirements.txt, djangorestframework
djangorestframework==3.11.1  # via -r requirements.txt
pytz==2020.1              # via -r requirements.txt, django
sqlparse==0.3.1           # via -r requirements.txt, django
